Move `Tests\LoggerPrinter` to `Helpers\FileLogger`
- Move `PhpCfdi\Finkok\Tests\LoggerPrinter` to `PhpCfdi\Finkok\Helpers\FileLogger` to be able to use it by default in the project.
