+++ docs/CHANGELOG.md
@@ -6,6 +6,13 @@ Nos apegamos a [SEMVER](SEMVER.md), revisa la información para entender mejor e
 
 Estos cambios se aplican y se publican, pero aún no son parte de una versión liberada.
 
+## Versión 0.5.5 2024-05-24
+
+Se mueve `PhpCfdi\Finkok\Tests\LoggerPrinter` a `PhpCfdi\Finkok\Helpers\FileLogger` para permitir la distribución
+de la herramienta dentro de la librería.
+
+
+
 ## Versión 0.5.4 2024-04-12
 
 Se actualiza el año de la licencia a 2024.
